graduate/pythonProject1/requirements.txt
fastapi
httpx
beautifulsoup4
apscheduler==3.10.4
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy-utils==0.41.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dateutil==2.8.2